Check for empty string before executing kill

+++ app/BE/stage_server/scripts/application-stop.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 pid=`ps aux | grep "node out/index.js" | grep -v grep | tr -s " " " " | cut -d " " -f2`
-kill -9 $pid
+# if string is not empty
+if [ -n $pid ]
+then
+    kill -9 $pid
+fi
